Made hasVisited methods of BruteForce and RouteFind, otherwise there's a duplication in the ofxTSP namespace.
RouteFinder.cpp
BruteForce.cpp

Changed the problem.getSymmetricCost call in both so that it creates a local variable - it wouldn't compile when creating that object within the function call - not sure why.

I also changed the #include files - for some reason it fails when I use the ofxTSP subgroup - not sure what's going on there - possible Xcode set up failure.